requirements.txt
# Include everything the framework requires
# You will automatically get updates for all versions starting with "1.".
rlbot==1.*

--find-links https://download.pytorch.org/whl/torch_stable.html
#torch>=1.11.0+cu115
torch>=1.11.0
rlgym-compat==1.*
rlgym-tools==1.*
numpy==1.*
stable_baselines3==1.*
pickle5>=0.0.11

